fix: user DTO 패키지 디렉토리 변경
