+++ client/user/src/main/java/com/logistic/client/user/presentation/requestDto/UpdateDeliveryManagerDTO.java
@@ -1,8 +1,6 @@
-package com.logistic.client.user.application.dto.requestDto;
+package com.logistic.client.user.presentation.requestDto;
 
-import com.logistic.client.user.domain.model.DeliveryManager;
 import com.logistic.client.user.domain.model.DeliveryManagerType;
-import com.logistic.client.user.domain.model.User;
 import jakarta.validation.constraints.AssertTrue;
 import jakarta.validation.constraints.NotBlank;
 import lombok.*;
@@ -14,10 +12,7 @@ import java.util.UUID;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class DeliveryManagerDTO {
-    @NotBlank(message = "유저 이름을 입력해주세요.")
-    private String username;
-
+public class UpdateDeliveryManagerDTO {
     private UUID hubId;
 
     @NotBlank(message = "배송 담당 타입을 입력해주세요.")
@@ -26,14 +21,5 @@ public class DeliveryManagerDTO {
     @AssertTrue(message = "COMPANY_DELIVERY_MANAGER 타입일 경우 허브 아이디는 필수 입력값입니다.")
     public boolean isHubIdValid() {
         return deliveryManagerType != DeliveryManagerType.COMPANY_DELIVERY_MANAGER || hubId != null;
-    }
-
-    public DeliveryManager toDeliveryManager(User user, int assignmentOrder) {
-        return DeliveryManager.builder()
-                .user(user)
-                .hubId(hubId)
-                .deliveryManagerType(deliveryManagerType)
-                .assignmentOrder(assignmentOrder)
-                .build();
     }
 }
